Testcase for adding backup rule
